Fixed bug regarding manufacturer input
If the input was 00000000, for example in the pokemon red rom, the
input field would appear blank, but it would still pass the check. This
resulted in a bad download. Now, if it reads 00000000 it won’t be
blank, but will say NULL. There may possibly be more issues converting
0000 to an ASCII character.